move scrollbar logic to desktop specific code
